requirements.txt
bcrypt
PyYAML